WIP: Add HTTPAgent
Added HTTPAgent which is a struct that is in charge of sending HTTP
requests and returning the response to the caller.

This is different from `Connection` which also sends HTTP requests, and
does JSON unmarshaling and other things based on the response code. The
latter responsibilities now fall upon the caller of HTTPAgent.

HTTPAgent is only being used by FindTeam() at this time. We added
`Deprecated` comments to `Connection` and related functions and fields.
Future editors of go-concourse should try and move over to using
HTTPAgent instead of `Connection` for making HTTP requests.

Signed-off-by: Izabela Gomes <igomes@pivotal.io>
Co-authored-by: Taylor Silva <tsilva@pivotal.io>